Task2

+++ Tyuiu.KupriyanovEA.Sprint3.Task2.V9.Lib/DataService.cs
@@ -10,7 +10,7 @@ namespace Tyuiu.KupriyanovEA.Sprint3.Task2.V9.Lib
 {
     public class DataService : ISprint3Task2V9
     {
-        public double GetSumSeries(double value, int startValue, int stopValue)
+        public double GetSumSeries(int value, int startValue, int stopValue)
         {
             double multSeries = 0;
             do
