+++ Tyuiu.KupriyanovEA.Sprint3.Task2.V9/Program.cs
@@ -29,7 +29,7 @@ namespace Tyuiu.KupriyanovEA.Sprint3.Task2.V9
             Console.WriteLine("* ИСХОДНЫЕ ДАННЫЕ:                                                        *");
             Console.WriteLine("***************************************************************************");
 
-            double value = 5;
+            int value = 5;
             int startValue = 1;
             int stopValue = 10;
